ai_service/main_simple.py
# AI Service - Enhanced Rule-Based Edition
from fastapi import FastAPI
from pydantic import BaseModel
import random
import json
from typing import List, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="Influencelytic AI Service", 
    description="Enhanced rule-based AI for influencer analytics",
    version="1.1.0"
)

@app.get("/health")
async def health():
    return {
        "status": "healthy",
        "service": "Influencelytic AI Service (Enhanced)",
        "version": "1.1.0",
        "ai_type": "rule_based_enhanced"
    }

@app.get("/")
async def root():
    return {"message": "Influencelytic AI Service - Ready!", "docs": "/docs"}

# Add your enhanced AI endpoints here...

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
