ai_service/requirements.txt
# Python 3.10.0 Compatible Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.4
pandas==2.0.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
tokenizers==0.15.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
httpx==0.25.2
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
pytest==7.4.3
black==23.11.0
rich==13.7.0
